Make <Shift>Q reformat a paragraph
This is much more useful then a nop. I found I would attempt to join a
whole paragraph to make it soft wrapped. Some blogs promote the function
should reformat the paragraph (gqip) but I found the composed command
makes sense to me and I'd use the composed version more often. While
the reverse (vipJ) is not very composability mainly because of the
non-standard behavior of the join (J) command.